fix: remove __init__ in main and build complete
